SPOTIFYFOLDER/spotifyProject/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
Django==4.2.5
django-channels==0.7.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Markdown==3.4.4
msgpack==1.0.5
oauthlib==3.2.2
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
